third_party/blink/web_tests/platform/generic/http/tests/misc/frame-default-enc-same-domain-expected.txt

CONSOLE MESSAGE: windows-1256
